chore: add shared utility modules
