+++ src/lib/text.ts
@@ -1,0 +1,9 @@
+export function matchText(q: string, label: string) {
+  return q
+    .trim()
+    .toLowerCase()
+    .split(/\s+/)
+    .filter(Boolean)
+    .every((p) => label.toLowerCase().includes(p));
+}
+
